+++ src/utils/flux/store/store.js
@@ -53,24 +53,21 @@ const CHANGE_EVENT = 'change';
  * @extends EventEmitter
  */
 export default class Store extends EventEmitter {
-
   constructor(name, data, opts = {}) {
     super(name, data, opts);
+
+    this.setMaxListeners(50);
 
     const suffix = `Check the initialization of ${this.constructor.name}.`;
 
     // tell the developer if they have not defined the name property.
     if (!name) {
-      throw new Error(
-        `You need to initialize your store with a name. ${suffix}`
-      );
+      throw new Error(`You need to initialize your store with a name. ${suffix}`);
     }
 
     // tell the developer if they have not defined the data property.
     if (!data) {
-      throw new Error(
-        `You need to initialize your store with data. ${suffix}`
-      );
+      throw new Error(`You need to initialize your store with data. ${suffix}`);
     }
 
     /**
@@ -215,5 +212,4 @@ export default class Store extends EventEmitter {
     // we use the store name so the view component knows which store updated
     this.emit(CHANGE_EVENT, this.name);
   }
-
 }